pyproject.toml
[tool.poetry]
name = "repl_python3_beta_Swiper-Bot-v3"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
youtube_dl = "^2021.6.6"
discord = "^1.7.3"
youtube-search = "^2.1.0"
validators = "^0.18.2"
PyNaCl = "^1.4.0"
google-api-python-client = "^2.22.0"
boto3 = "^1.19.12"
google-cloud-vision = "^2.6.2"
opencv-python-headless = "^4.5.4"
imageio = "^2.10.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
